Manage keys and keyrings

This set of functions handles keys that are needed to download products such as satellite data.

Usage

create_keyring(ringname, password)

delete_keyring(ringname)

log_in(ringname, password)

log_out(ringname)

add_key(ringname, provider, username, password)

get_username(ringname, provider)

get_password(ringname, provider)

delete_key(ringname, provider)

check_keyring(ringname)

Arguments

ringname

character. The name of a keyring (a ringname).

password

character. A password.

provider

character. The name of a provider (e.g. "modis").

username

character. A username signed-up in a service provider.

Value

nothing.

Details

This set of function are wrappers of the keyring package. A keyring is an account that can hold keys (credentials) for various services. Once a keyring is created, the user can start adding keys to it. Functions in the cronus package that require credentials can automatically get them if provided with the ringname. Users are reminded to log-in and log-out in order to use their keyrings.

Functions

create_keyring(): Create a keyring.

delete_keyring(): Delete a keyring.

log_in(): Log into your keyring.

log_out(): Log out of your keyring.

add_key(): Add a key to your keyring.

get_username(): Get a username.

get_password(): Get a password.

delete_key(): Delete a key.

check_keyring(): Check whether the user is logged-in.

Examples

if (FALSE) {
ringname <- "my_ringname"
password <- "my_password"

create_keyring(ringname, password)
check_keyring(ringname)

log_in(ringname, password)
check_keyring(ringname)

add_key(ringname = ringname,
provider = "nass",
username = NULL,
password = "nass_key")

add_key(ringname = ringname,
        provider = "usgs",
        username = "usgs_username",
        password = "usgs_password")

get_username(ringname, "usgs")
get_password(ringname, "usgs")
delete_key(ringname, "usgs")

log_out(ringname)
delete_keyring(ringname)
}